gcp-pubsub-examples/python-publisher/requirements.txt
google-cloud-pubsub==2.18.4
google-auth==2.23.4
